+++ Changes.md
@@ -29,6 +29,7 @@ Fixes
 - Cryptomatte :
   - Fixed errors when the input image didn't contain the main `RGBA` channels.
   - Fixed inaccurate hash.
+- SceneAlgo : Fixed exception handling for Python render adaptors. Previously an exception during adaptor construction caused a `SystemError`.
 
 API
 ---